chore: Migrate to CRXJS completed
