--- .github/workflows/submit.yml
@@ -1,91 +1,0 @@
-name: Submit for Review
-on:
-  workflow_dispatch:
-    inputs:
-      dryRun:
-        default: false
-        type: boolean
-        description: Skip submission and perform a dry run
-
-jobs:
-  validate:
-    name: Validate
-    uses: ./.github/workflows/validate.yml
-    secrets: inherit
-
-  submit:
-    name: Submit
-    runs-on: ubuntu-22.04
-    needs: [validate]
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - uses: oven-sh/setup-bun@v2
-
-      - name: Configure Git
-        run: |
-          git config user.email "github-actions@users.noreply.github.com"
-          git config user.name "GitHub Actions"
-
-      - name: Install dependencies
-        run: bun install --frozen-lockfile
-
-      - name: Bump Version
-        id: version
-        run: |
-          bunx changelogen@latest --bump
-          NEW_VERSION=$(bun --print 'require("./package.json").version')
-          echo "newVersion=$NEW_VERSION" >> "$GITHUB_OUTPUT"
-
-      - name: Build and Zip
-        run: |
-          bun run zip
-          bun run zip:firefox
-
-      - name: Upload Artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          path: .output/*.zip
-          if-no-files-found: error
-          include-hidden-files: true
-
-      - name: Submit
-        run: |
-          bun wxt submit \
-            --chrome-zip .output/*-chrome.zip \
-            --firefox-zip .output/*-firefox.zip --firefox-sources-zip .output/*-sources.zip
-        env:
-          DRY_RUN: ${{ inputs.dryRun }}
-          CHROME_EXTENSION_ID: ${{ secrets.CHROME_EXTENSION_ID }}
-          CHROME_CLIENT_ID: ${{ secrets.CHROME_CLIENT_ID }}
-          CHROME_CLIENT_SECRET: ${{ secrets.CHROME_CLIENT_SECRET }}
-          CHROME_REFRESH_TOKEN: ${{ secrets.CHROME_REFRESH_TOKEN }}
-          FIREFOX_EXTENSION_ID: ${{ secrets.FIREFOX_EXTENSION_ID }}
-          FIREFOX_JWT_ISSUER: ${{ secrets.FIREFOX_JWT_ISSUER }}
-          FIREFOX_JWT_SECRET: ${{ secrets.FIREFOX_JWT_SECRET }}
-          EDGE_PRODUCT_ID: ${{ secrets.EDGE_PRODUCT_ID }}
-          EDGE_CLIENT_ID: ${{ secrets.EDGE_CLIENT_ID }}
-          EDGE_CLIENT_SECRET: ${{ secrets.EDGE_CLIENT_SECRET }}
-          EDGE_ACCESS_TOKEN_URL: ${{ secrets.EDGE_ACCESS_TOKEN_URL }}
-
-      - name: Commit and Push
-        if: ${{ !inputs.dryRun }}
-        run: |
-          git add package.json CHANGELOG.md
-          git commit -am "chore(release): v$NEW_VERSION"
-          git tag v$NEW_VERSION
-          git push
-          git push --tags
-        env:
-          NEW_VERSION: ${{ steps.version.outputs.newVersion }}
-
-      - name: Release
-        if: ${{ !inputs.dryRun }}
-        run: pnpx changelogen@latest gh release ${{ steps.version.outputs.newVersion }} --token ${{ github.token }}
-
-      # Don't upload sources zip - it can contain .env files, which may include secrets
-      - name: Upload ZIPs
-        if: ${{ !inputs.dryRun }}
-        run: gh release upload ${{ steps.version.outputs.newVersion }} .output/*-chrome.zip .output/*-firefox.zip
