+++ .gitignore
@@ -8,7 +8,7 @@ pnpm-debug.log*
 lerna-debug.log*
 
 node_modules
-.output
+dist
 artifacts
 *.local
 coverage
