+++ .prettierignore
@@ -3,6 +3,6 @@ dist-*
 artifacts
 coverage
 .wxt
-.output
+dist
 coverage
 *.lock
