+++ package.json
@@ -6,22 +6,18 @@
   "type": "module",
   "scripts": {
     "check": "check",
-    "dev": "wxt",
-    "dev:firefox": "wxt --browser firefox",
-    "build": "wxt build",
-    "build:firefox": "wxt build --browser firefox",
+    "dev": "vite",
+    "build": "vite build",
     "test": "vitest",
     "test:coverage": "vitest run --coverage",
-    "zip": "wxt zip",
-    "zip:firefox": "wxt zip -b firefox",
-    "prepare": "simple-git-hooks",
-    "postinstall": "wxt prepare"
+    "prepare": "simple-git-hooks"
   },
   "dependencies": {
     "@aklinker1/check": "^1.4.5",
     "@tanstack/vue-query": "^5.59.13",
     "@webext-core/messaging": "^2.0.2",
     "@webext-core/proxy-service": "^1.2.0",
+    "@wxt-dev/storage": "^1.1.0",
     "async-mutex": "^0.5.0",
     "fast-deep-equal": "^3.1.3",
     "minimatch": "^10.0.1",
@@ -29,15 +25,15 @@
   },
   "devDependencies": {
     "@aklinker1/generate-changelog": "^1.1.2",
+    "@crxjs/vite-plugin": "^2.0.0-beta.31",
     "@iconify/json": "^2.2.261",
     "@tailwindcss/vite": "^4.0.6",
     "@types/bun": "^1.1.11",
-    "@types/chrome": "^0.0.278",
+    "@types/chrome": "^0.0.304",
     "@types/jsdom": "^21.1.7",
+    "@vitejs/plugin-vue": "^5.2.1",
     "@vitest/coverage-v8": "^2.1.3",
-    "@wxt-dev/auto-icons": "^1.0.2",
     "@wxt-dev/i18n": "^0.2.1",
-    "@wxt-dev/module-vue": "^1.0.1",
     "daisyui": "^5.0.0-beta.7",
     "env-cmd": "^10.1.0",
     "fast-glob": "^3.3.2",
@@ -49,10 +45,10 @@
     "tailwindcss": "^4.0.6",
     "typescript": "^5.6.3",
     "unplugin-icons": "^0.19.3",
+    "vite": "6.0.8",
     "vitest": "^2.1.3",
     "vue": "^3.5.12",
-    "vue-tsc": "^2.1.6",
-    "wxt": "^0.19.11"
+    "vue-tsc": "^2.1.6"
   },
   "simple-git-hooks": {
     "pre-commit": "bun lint-staged"
