+++ src/components/CommitDiff.ts
@@ -1,3 +1,6 @@
+import { i18n } from "@/utils/i18n";
+import { createDiffComponent } from "./createDiffComponent";
+
 export const CommitDiff = createDiffComponent({
   getAdditionsElement: () =>
     document.querySelector<HTMLElement>("#toc>*>strong:nth-child(2)"),
