+++ src/components/CustomListItem.vue
@@ -1,4 +1,6 @@
 <script lang="ts" setup>
+import { i18n } from "@/utils/i18n";
+
 defineProps<{
   value: string;
 }>();
