+++ src/components/ShowGeneratedCountPref.vue
@@ -1,4 +1,6 @@
 <script lang="ts" setup>
+import { i18n } from "@/utils/i18n";
+
 const hideLineCount = defineModel<boolean>("hideGeneratedLineCount", {
   required: true,
 });
