+++ src/composables/useGithubUserQuery.ts
@@ -1,4 +1,6 @@
 import { getGithubService, Github } from "@/utils/github";
+import { QueryKeys } from "@/utils/QueryKeys";
+import { useQuery } from "@tanstack/vue-query";
 
 export default function (token: { value: string | undefined }) {
   const github = getGithubService();
