--- src/entrypoints/background.ts
@@ -1,12 +1,0 @@
-import { registerGithubService, createGithubApi } from "@/utils/github";
-
-export default defineBackground(() => {
-  const api = createGithubApi();
-  registerGithubService(api);
-
-  browser.runtime.onInstalled.addListener(async ({ reason }) => {
-    if (reason === "install") {
-      void browser.runtime.openOptionsPage();
-    }
-  });
-});
