+++ src/options.html
@@ -4,9 +4,8 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./main.ts" type="module"></script>
+    <script src="./options.ts" type="module"></script>
     <link rel="stylesheet" href="~/assets/tailwind.css" />
-    <meta name="manifest.open_in_tab" content="false" />
   </head>
   <body class="p-4 gap-4 flex flex-col"></body>
 </html>
