+++ src/options.ts
@@ -1,5 +1,7 @@
 import Options from "@/pages/Options.vue";
 import { VueQueryPlugin } from "@tanstack/vue-query";
+import { i18n } from "./utils/i18n";
+import { createApp } from "vue";
 
 document.title = i18n.t("options.title");
