+++ src/utils/cache/createKeyValueCache.ts
@@ -1,3 +1,4 @@
+import { storage } from "@wxt-dev/storage";
 import { Mutex } from "async-mutex";
 
 interface CachedValue<TValue> {
