+++ src/utils/i18n.ts
@@ -1,0 +1,3 @@
+import { createI18n } from "@wxt-dev/i18n";
+
+export const i18n = createI18n();
